+++ routes/web/logout.js
@@ -1,0 +1,6 @@
+module.exports = function(app) {
+	app.get('/logout', function(req, res) {
+	  	req.logout();
+	  	res.redirect('/');
+	});
+}
